refactor: `ExtensionDialog` 이동
