+++ packages/shared/src/modules/extension-bridge/ExtensionBridge.ts
@@ -192,7 +192,6 @@ export default class ExtensionBridge {
     try {
       chrome.runtime.sendMessage(EXTENSION.id, { type: BRIDGE_MESSAGE_TYPES.GET_EXTENSION_MANIFEST }, callbackFn);
     } catch (error) {
-      console.log('여기서 에러가 발생했습니다.');
       throw new ExtensionError('Failed to request extension manifest', ExtensionErrorCode.RUNTIME_ERROR, error);
     }
   }
