+++ packages/web/src/app/[lng]/(auth)/memos/page.tsx
@@ -11,6 +11,7 @@ import { Suspense } from 'react';
 
 import { HeaderMargin } from '../../_components/Header';
 import { MemoSearchForm, MemoSearchFormProvider, MemoView } from './_components';
+import ExtensionDialog from './_components/ExtensionDialog';
 
 interface PageProps extends LanguageParams {
   searchParams: SearchParamsType;
@@ -30,6 +31,8 @@ export default async function Page({ searchParams, params: { lng } }: PageProps)
           </MemoSearchFormProvider>
         </Suspense>
       </HydrationBoundaryWrapper>
+
+      <ExtensionDialog lng={lng} />
     </main>
   );
 }
