+++ packages/web/src/app/_components/index.ts
@@ -1,4 +1,4 @@
-export { default as ExtensionDialog } from './ExtensionDialog';
+export { default as ExtensionDialog } from '../[lng]/(auth)/memos/_components/ExtensionDialog';
 export { default as Introduction } from './Introduction';
 export { default as WebVitals } from './WebVitals';
 export { default as InitDayjs } from './InitDayjs';
